playlist manager: added WebStream support
