+++ src/main/java/net/pms/store/utils/StreamMeta.java
@@ -1,0 +1,12 @@
+package net.pms.store.utils;
+
+
+public class StreamMeta {
+
+	public String logo;
+	public String genre;
+	public String contentType;
+	public Integer sample;
+	public Integer bitrate;
+
+}
